% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rice.bg}
\alias{rice.bg}
\title{Rice Ballgown Object}
\format{a ballgown object: 33 transcripts and 6 samples}
\usage{
rice.bg
}
\description{
Small ballgown object created with a subset of rice RNAseq data,
for demonstration purposes
}
\examples{
data(rice.bg)
rice.bg
# ballgown instance with 33 transcripts and 6 samples
}
\keyword{datasets}
