% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splicePlot.R
\name{splicePlot}
\alias{splicePlot}
\title{Gene Splicing Plot}
\usage{
splicePlot(bg, gene, samples, bam.dir = NA, start = NA, end = NA, 
            labels = samples, junc.type = c("score", "count"), junc.text = TRUE,
            trans.select = "rowMaxs(x)>=1", junc.select = "rowMaxs(x)>=5",
            col = SushiColors(2)(length(samples) + 1)[-1],transparency = 0.5,
            scale = "Kb", plotgenetype = "arrow", ...)
}
\arguments{
\item{bg}{ballgown object. See \code{\link[=ballgown-class]{ballgown}}.}

\item{gene}{string indicating a gene ID (must be in the 'bg')}

\item{samples}{names of the samples to be shown (must be in the 'bg' and
have bam files in the 'bam.dir')}

\item{bam.dir}{bam file directory of the samples.
If NA, instead of read depth, conserved exons are drawn.}

\item{start}{start position to be shown. If NA, start position of the gene
will be used.}

\item{end}{stop position to be shown. If NA, end position of the gene will
be used.}

\item{labels}{labels for samples (default: sample names).
If it is NA, neigher sample names nor gene names will be labeled}

\item{junc.type}{type of junction estimates to be shown ('score' for
junction score; count' for junction read count)}

\item{junc.text}{TRUE/FALSE indicating whether junction estimates should
be labeled}

\item{trans.select}{logical expression-like string, indicating transcript
rows to select from a matrix of transcript coverages: NA value keeps all
transcripts. See \code{\link{spliceGene}}}

\item{junc.select}{logical expression-like string, indicating junction rows
to select from a matrix of junction counts: NA value keeps all junctions.
See \code{\link{spliceGene}}}

\item{col}{a vector of length(samples) specifying colors of read depths.}

\item{transparency}{value between 0 and 1 indicating the degree of 
transparency of read depths.}

\item{scale}{scale of the \code{\link[Sushi]{labelgenome}} ('bp','Kb','Mb')}

\item{plotgenetype}{string specifying whether the genes should resemble a
'box' or a 'arrow'. See \code{\link[Sushi]{plotGenes}}.}

\item{...}{values to be passed to \code{\link[Sushi]{plotGenes}}.}
}
\value{
see \code{\link[Sushi]{plotGenes}}.
}
\description{
Visualization of read coverage, splicing information and gene information in
a gene region. This function is a wrapper of \code{\link{getDepth}},
\code{\link{getGeneinfo}}, \code{\link{spliceGene}},
\code{\link[Sushi]{plotBedgraph}} and \code{\link[Sushi]{plotGenes}}.
}
\examples{

data(rice.bg)
rice.bg

samples <- paste('Sample', c('027','102','237'),sep='_')
bam.dir <- system.file('extdata',package = 'vasp')

## plot the whole gene region
splicePlot(rice.bg,samples,bam.dir,gene='MSTRG.183',bheight=0.2)

## plot the alternative splicing region
splicePlot(rice.bg,samples,bam.dir,gene='MSTRG.183',start=1179000)
}
